Add preliminary uprobe support
RegisterUprobe() is added to EventMonitor to allow registration of uprobe
events. Symbol resolution is done, but there is no caching of any kind.
Uprobes won't work as expected with some union filesystems. More work needs
to be done here, but the support that's here is functional for system-wide
host monitoring, which is all that it's intended for at this time.